# backend/internal/delivery/http/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: BRAG 2026 API
  version: "1.0"
  description: |
    API untuk BNI Grow Annual Challenge — mesin skor, leaderboard, dan
    manajemen musim kompetisi.

    ## Otentikasi

    Ada dua cara, dan keduanya memakai izin yang sama persis. Rute yang
    butuh admin tetap butuh admin, dipanggil dari browser maupun dari skrip.

    **Cookie sesi** — untuk aplikasi web. `POST /api/auth/login` menaruh
    cookie `HttpOnly`; sertakan pada tiap permintaan berikutnya.

    **Kunci API** — untuk integrasi. Kirim lewat salah satu header:

    ```
    X-API-Key: brag_xxxxxxxxxxxx
    Authorization: Bearer brag_xxxxxxxxxxxx
    ```

    Kunci bertindak atas nama seorang pengguna dan mewarisi perannya. Kunci
    bertanda *read-only* hanya melayani `GET`; permintaan tulis dijawab
    `403`. Kunci ditampilkan **sekali** saat dibuat dan tidak bisa dilihat
    lagi — hanya sidik jarinya yang disimpan.

    ## Kesalahan

    Kegagalan selalu berbentuk `{ "error": "pesan untuk pengguna" }` dengan
    kode status yang sesuai: `400` input tidak valid, `401` belum masuk,
    `403` peran tidak mencukupi, `404` tidak ditemukan, `409` bentrok
    dengan keadaan sekarang, `429` terlalu sering.

    ## Paginasi

    Daftar panjang menerima `limit` (maks 200) dan `offset`, atau `page`.
    Jawabannya membawa `items`, `total`, `limit`, dan `offset`.
servers:
  - url: http://localhost:8081
    description: Pengembangan lokal
...
security:
  - apiKeyHeader: []
  - bearerKey: []
  - sessionCookie: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Masuk dengan email dan kata sandi
      description: |
        Dibatasi 10 percobaan per menit per alamat IP. Balasan menyetel
        cookie sesi `HttpOnly`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: demo.admin@brag2026.id }
                password: { type: string, format: password, example: demo123 }
      responses:
        "200":
          description: Berhasil masuk
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Identitas pemanggil beserta profil kompetisinya
      responses:
        "200":
          description: Pengguna saat ini
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  member:
                    oneOf:
                      - { $ref: "#/components/schemas/Member" }
                      - { type: "null" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Ganti kata sandi sendiri
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/api-keys:
    get:
      tags: [API Keys]
      summary: Daftar kunci
      description: Admin melihat seluruh kunci; peran lain hanya miliknya sendiri.
      responses:
        "200":
          description: Daftar kunci
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/APIKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [API Keys]
      summary: Buat kunci baru
      description: |
        Nilai kunci hanya muncul di balasan ini dan tidak dapat diambil lagi.
        Menerbitkan kunci atas nama orang lain hanya boleh dilakukan admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nama]
              properties:
                nama:
                  type: string
                  description: Label agar kunci mudah dikenali
                  example: Integrasi dashboard
                user_id:
                  type: string
                  format: uuid
                  description: Pemilik kunci. Kosong berarti pemanggil sendiri.
                read_only:
                  type: boolean
                  default: true
                  description: Bila true, kunci hanya melayani GET.
                expires_in_days:
                  type: integer
                  minimum: 0
                  maximum: 3650
                  description: 0 berarti tidak kedaluwarsa sendiri.
      responses:
        "201":
          description: Kunci dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: Rahasianya. Simpan sekarang — tidak bisa dilihat lagi.
                    example: brag_rA6bOQvZwhJRqIxK2mN8pQ7sT4vW1yZ0
                  record: { $ref: "#/components/schemas/APIKey" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Cabut kunci
      description: Berlaku seketika. Tidak bisa dibatalkan.
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Kunci sudah dicabut sebelumnya
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/public/leaderboard:
    get:
      tags: [Publik, Leaderboard]
      summary: Klasemen tim tanpa otentikasi
      security: []
      parameters: [{ $ref: "#/components/parameters/Kategori" }]
      responses:
        "200": { $ref: "#/components/responses/Leaderboard" }
  /api/public/leaderboard/teams/{id}/history:
    get:
      tags: [Publik, Leaderboard]
      summary: Riwayat poin sebuah tim
      security: []
      parameters:
        - { $ref: "#/components/parameters/PathID" }
        - { $ref: "#/components/parameters/KategoriLedger" }
      responses:
        "200": { $ref: "#/components/responses/LedgerList" }
  /api/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Klasemen tim dan individu
      description: |
        Daftar individu dipotong oleh `limit`, jadi `kategori` menentukan
        urutannya di server — mengurutkan ulang di klien akan memeringkat
        potongan yang salah.
      parameters:
        - { $ref: "#/components/parameters/Kategori" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { $ref: "#/components/responses/Leaderboard" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/leaderboard/teams/{id}/history:
    get:
      tags: [Leaderboard]
      summary: Riwayat poin sebuah tim
      parameters:
        - { $ref: "#/components/parameters/PathID" }
        - { $ref: "#/components/parameters/KategoriLedger" }
      responses:
        "200": { $ref: "#/components/responses/LedgerList" }
  /api/dashboard:
    get:
      tags: [Member]
      summary: Ringkasan musim untuk pemanggil
      responses:
        "200":
          description: Skor, peringkat, badge, dan event minggu ini
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dashboard" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/members/search:
    get:
      tags: [Member]
      summary: Cari anggota untuk pemilih lawan transaksi
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: Nama atau email
      responses:
        "200":
          description: Anggota yang cocok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Member" }
  /api/tyfcb:
    post:
      tags: [Member]
      summary: Catat transaksi TYFCB
      description: |
        Masuk sebagai `pending`. Poin baru menyentuh ledger setelah admin
        memverifikasi. Skornya dihitung saat pencatatan:
        `round(Band × PairPenalty × EventMultiplier)`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TyfcbInput" }
      responses:
        "201":
          description: Transaksi tercatat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TyfcbEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/visitors:
    post:
      tags: [Member]
      summary: Daftarkan tamu undangan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VisitorInput" }
      responses:
        "201":
          description: Tamu terdaftar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visitor" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Kontak sudah terdaftar di musim ini
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/activity:
    get:
      tags: [Katalog]
      summary: Umpan aktivitas musim
      parameters: [{ $ref: "#/components/parameters/Limit" }]
      responses:
        "200":
          description: Kejadian terbaru
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityItem" }
...
  /api/boosters:
    get:
      tags: [Katalog]
      summary: Pengumuman booster
      parameters:
        - name: active
          in: query
          schema: { type: boolean }
          description: Bila true, hanya yang sedang berjalan
      responses:
        "200":
          description: Booster
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booster" }
  /api/boosters/{id}:
    get:
      tags: [Katalog]
      summary: Satu booster
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200":
          description: Booster
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booster" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/prizes:
    get:
      tags: [Katalog]
      summary: Prize pool
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected, awarded] }
      responses:
        "200":
          description: Hadiah
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Prize" }
  /api/prizes/donate:
    post:
      tags: [Member]
      summary: Donasikan hadiah
      description: Masuk sebagai `pending` hingga admin menyetujui, lalu donatur memperoleh badge PATRON.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PrizeInput" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/one-to-one:
    get:
      tags: [Member]
      summary: Catatan pertemuan 1-2-1
      parameters: [{ $ref: "#/components/parameters/Limit" }]
      responses:
        "200":
          description: Catatan
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OneToOne" }
    post:
      tags: [Member]
      summary: Catat pertemuan 1-2-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [member_id, tanggal]
              properties:
                member_id: { type: string, format: uuid }
                tanggal: { type: string, format: date }
                catatan: { type: string, nullable: true }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/tour/voice:
    get:
      tags: [Publik]
      summary: Narasi suara satu langkah
      description: Mengembalikan `204` bila layanan text-to-speech tidak dikonfigurasi.
      security: []
      parameters:
        - name: step
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Audio narasi
          content:
            audio/mpeg:
              schema: { type: string, format: binary }
        "204":
          description: Tidak dikonfigurasi — klien memakai suara peramban
  /api/export/leaderboard:
    get:
      tags: [Export]
      summary: Unduh leaderboard
      description: |
        Excel menyimpan angka sebagai angka dengan format tampilan, jadi
        kolomnya bisa langsung dijumlah.
      parameters: [{ $ref: "#/components/parameters/Format" }]
      responses:
        "200": { $ref: "#/components/responses/FileDownload" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/captain/team:
    get:
      tags: [Kapten]
      summary: Anggota tim yang dipimpin
      responses:
        "200":
          description: Tim beserta anggotanya
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaptainTeam" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/captain/tyfcb:
    post:
      tags: [Kapten]
      summary: Catat TYFCB atas nama anggota tim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/TyfcbInput" }
                - type: object
                  required: [member_id]
                  properties:
                    member_id:
                      type: string
                      format: uuid
                      description: Penjual. Harus satu tim dengan kapten.
      responses:
        "201":
          description: Transaksi tercatat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TyfcbEntry" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/captain/tyfcb/{id}/void:
    patch:
      tags: [Kapten]
      summary: Batalkan transaksi
      description: Bila sudah terverifikasi, poinnya dibalik dengan baris ledger bertanda negatif.
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/captain/visitors:
    post:
      tags: [Kapten]
      summary: Daftarkan tamu atas nama anggota tim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/VisitorInput" }
                - type: object
                  required: [member_id]
                  properties:
                    member_id: { type: string, format: uuid }
      responses:
        "201":
          description: Tamu terdaftar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visitor" }
  /api/captain/visitors/{id}/void:
    patch:
      tags: [Kapten]
      summary: Batalkan tamu dan balikkan poinnya
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/captain/members/{id}/password:
    patch:
      tags: [Kapten]
      summary: Setel ulang kata sandi anggota tim
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/members:
    get:
      tags: [Admin]
      summary: Daftar anggota
      parameters:
        - { $ref: "#/components/parameters/Search" }
        - { $ref: "#/components/parameters/TeamID" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200":
          description: Halaman anggota
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Page" }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Member" }
    post:
      tags: [Admin]
      summary: Tambah anggota
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberInput" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/admin/members/{id}:
    patch:
      tags: [Admin]
      summary: Ubah anggota
      description: Menaikkan `color_status` memicu bonus Level Up dan badge LEVEL_UP.
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberInput" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /api/admin/tyfcb:
    get:
      tags: [Admin]
      summary: Antrean verifikasi TYFCB
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, verified, rejected, void] }
        - { $ref: "#/components/parameters/Search" }
        - { $ref: "#/components/parameters/TeamID" }
        - { $ref: "#/components/parameters/DateFrom" }
        - { $ref: "#/components/parameters/DateTo" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200":
          description: Halaman transaksi beserta jumlah per status
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/TyfcbEntry" }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  counts:
                    type: object
                    additionalProperties: { type: integer }
  /api/admin/tyfcb/{id}:
    patch:
      tags: [Admin]
      summary: Setujui atau tolak transaksi
      description: |
        Menjadi `verified` mengkredit poin; keluar dari `verified` menulis
        baris pembalik yang sama besar. Penolakan **wajib** menyertakan
        alasan — anggota membacanya di riwayatnya sendiri.
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [pending, verified, rejected] }
                reason:
                  type: string
                  nullable: true
                  description: Wajib bila status `rejected`.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/admin/visitors:
    get:
      tags: [Admin]
      summary: Daftar tamu
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [terdaftar, hadir, hadir_penuh] }
        - name: converted
          in: query
          schema: { type: boolean }
        - { $ref: "#/components/parameters/Search" }
        - { $ref: "#/components/parameters/TeamID" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200":
          description: Halaman tamu
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Page" }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Visitor" }
  /api/admin/visitors/{id}:
    patch:
      tags: [Admin]
      summary: Perbarui kehadiran atau konversi tamu
      description: |
        Poin bersifat kumulatif — hadir 20, hadir penuh 50, konversi +100.
        Menurunkan status membalik selisihnya secara tepat.
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status_hadir: { type: string, enum: [terdaftar, hadir, hadir_penuh] }
                is_converted: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/admin/teams:
    get:
      tags: [Admin]
      summary: Daftar tim
      responses:
        "200":
          description: Tim
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Team" }
    post:
      tags: [Admin]
      summary: Buat tim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nama_tim]
              properties:
                nama_tim: { type: string }
      responses:
        "201": { $ref: "#/components/responses/Created" }
  /api/admin/teams/{id}:
    patch:
      tags: [Admin]
      summary: Ganti nama tim
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_tim: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
    delete:
      tags: [Admin]
      summary: Hapus tim
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/admin/classifications:
    get:
      tags: [Admin]
      summary: Daftar klasifikasi
      responses:
        "200":
          description: Klasifikasi
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Classification" }
    post:
      tags: [Admin]
      summary: Tambah klasifikasi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nama]
              properties: { nama: { type: string } }
      responses:
        "201": { $ref: "#/components/responses/Created" }
  /api/admin/classifications/{id}:
    patch:
      tags: [Admin]
      summary: Ganti nama klasifikasi
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { nama: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
    delete:
      tags: [Admin]
      summary: Hapus klasifikasi
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /api/admin/booster:
    get:
      tags: [Admin]
      summary: Daftar booster
      responses:
        "200":
          description: Booster
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booster" }
    post:
      tags: [Admin]
      summary: Buat booster
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BoosterInput" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
  /api/admin/booster/{id}:
    patch:
      tags: [Admin]
      summary: Ubah booster
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BoosterInput" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
    delete:
      tags: [Admin]
      summary: Hapus booster
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /api/admin/events/{id}:
    delete:
      tags: [Admin]
      summary: Hapus event terjadwal
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /api/admin/prizes:
    get:
      tags: [Admin]
      summary: Prize pool lengkap
      responses:
        "200":
          description: Hadiah
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Prize" }
    post:
      tags: [Admin]
      summary: Tambahkan hadiah panitia
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PrizeInput" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
...
  /api/admin/spheres/{id}:
    delete:
      tags: [Admin]
      summary: Hapus contact sphere
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /api/admin/one-to-one/{id}:
    delete:
      tags: [Admin]
      summary: Hapus catatan 1-2-1
      parameters: [{ $ref: "#/components/parameters/PathID" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
components:
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Kunci API pada header khusus.
    bearerKey:
      type: http
      scheme: bearer
      description: Kunci API yang sama, dikirim sebagai bearer token.
    sessionCookie:
      type: apiKey
      in: cookie
      name: brag_session
      description: Cookie sesi dari `POST /api/auth/login`.
...